Move basic types to shared crate
